更换为Kryox反序列化器,序列化到Redis

+++ src/main/java/cn/sola97/vrchat/service/MessageService.java
@@ -23,5 +23,5 @@ public interface MessageService {
 
     Long enqueueMessages(List<MessageDTO> queue);
 
-    int getMessageQueueSize();
+    Long getMessageQueueSize();
 }
